requirements.txt
anyio==4.12.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
playwright==1.57.0
proto-plus==1.27.0
protobuf==6.33.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
pyparsing==3.3.1
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
soupsieve==2.8.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
